Jumping and collision detection now works properly, but I broke requests from client while I was trying to fix them.

+++ 1v1_neckbeard_server/Entity.hpp
@@ -28,7 +28,7 @@ struct entity {
     
     direction dir;
     char hp;
-    sf::Vector2i pos;
+    sf::Vector2f pos;
     
 };
